interoperability_adapter/requirements.txt
Flask~=3.0.0
fhir.resources==5.0.1
psycopg2-binary~=2.9.9
